+++ src/components/domain/ManageGroup/UpdateGroupInfo.jsx
@@ -1,8 +1,12 @@
 import { Button, Heading, Input, TagInput, Text, Textarea } from '@/components/base';
+
 import { useGroupContext } from '@/context/GroupProvider';
-import { useToastContext } from '@/context/ToastProvider';
+
 import useInput from '@/hooks/useInput';
+import useToasts from '@/hooks/useToasts';
+
 import { useEffect, useState } from 'react';
+
 import { StyledGroupBox, StyledGroupInfo } from './styles';
 
 const ALERT_MESSAGE = {
@@ -16,7 +20,7 @@ const ALERT_MESSAGE = {
 function UpdateGroupInfo({ group, setGroup }) {
   const { headCount: memberCount, tagList: tags, intro: introduction } = group.description;
   const { onUpdateGroup } = useGroupContext();
-  const { addToast } = useToastContext();
+  const { addToast } = useToasts();
   const groupName = useInput('');
   const headCount = useInput(0);
   const intro = useInput('');
