+++ src/pages/SignIn.jsx
@@ -1,15 +1,21 @@
 import { postUserSignIn } from '@/api/userSign';
+
 import { imgLogin } from '@/assets/images';
+
 import { Button, Heading, Image, Spinner, Text } from '@/components/base';
 import SignInput from '@/components/domain/SignInput';
+
 import { useAuthContext } from '@/context/AuthProvider';
-import { useToastContext } from '@/context/ToastProvider';
-import { COLOR } from '@/styles/color';
-import theme from '@/styles/theme';
-import { css } from '@emotion/react';
-import styled from '@emotion/styled';
+
+import useToasts from '@/hooks/useToasts';
+
 import { useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+
+import { COLOR } from '@/styles/color';
+import { css } from '@emotion/react';
+import theme from '@/styles/theme';
+import styled from '@emotion/styled';
 
 const ERRORS = {
   EMAIL_EMPTY_ERROR: '이메일을 입력해 주세요.',
@@ -26,7 +32,7 @@ const INPUT_NUMBER_LIMIT = {
 function SignIn() {
   const navigate = useNavigate();
   const { onLogin } = useAuthContext();
-  const { addToast } = useToastContext();
+  const { addToast } = useToasts();
 
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
