+++ src/stores/toasts.js
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const toastsState = atom({
+  key: 'toasts',
+  default: [],
+});
